Refactored NavigationButtonView
