+++ afrigoal-afcon-2025/Services/TeamAPIService.swift
