+++ afrigoal-afcon-2025/Views/LoadingOverlay.swift
@@ -34,3 +34,8 @@ struct LoadingOverlay : View {
     }
 
 }
+
+#Preview {
+    
+    LoadingOverlay()
+}
