+++ afrigoal-afcon-2025/Views/SelectTeams/SelectTeamsHeader.swift
