+++ afrigoal-afcon-2025/Views/SelectTeams/TeamIcon.swift
